Explicitly sets the ARRAYSIZE as integer.
The host compiler treats the return value of ARRAYSIZE as unsigned
integer while the ARM cross-compiler does it differently. Use
explicit cast to make it more consistent.

+++ logupload/client/upload.c
@@ -209,7 +209,7 @@ int upload_file(const char *server_url, const char* target_name, char* data,
     return -1;
   }
 
-  for (i=0; i < ARRAYSIZE(resolvers); ++i) {
+  for (i=0; i < (int)ARRAYSIZE(resolvers); ++i) {
     long http_code;
     struct kvpair* kvpairs = kvpairs_save;
